pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advanced-research"
version = "0.1.0"
description = "Advanced Research framework for symbiotic intelligence with NeuralBlitz integration"
authors = [
    {name = "Advanced Research Team", email = "research@neuralblitz.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "jupyter>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    "markdown>=3.4.0",
    "pygments>=2.10.0",
]

[project.optional-dependencies]
lrs-agents = [
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.2.0",
    "httpx>=0.24.0",
]
opencode = [
    "opencode-api>=1.0.0",
    "gitpython>=3.1.0",
    "tree-sitter>=0.20.0",
]
neuralblitz = [
    "neuralblitz-v50>=1.0.0",
    "jax>=0.4.0",
    "flax>=0.7.0",
    "equinox>=0.10.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
all = [
    "advanced-research[lrs-agents,opencode,neuralblitz,dev]"
]

[project.scripts]
advanced-research = "advanced_research.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"